feat: add generic input field component
