+++ src/frontend/components/UtilityComponents.jsx
@@ -1,5 +1,6 @@
-import React from "react";
+import React, {useEffect, useState} from "react";
 import { observer } from "mobx-react-lite";
+import {preventNonDigitInput} from "../utils/utils.js";
 
 export const ChangingButton = observer(({ id, onClick, disabled, label }) => (
     <button
@@ -11,3 +12,44 @@ export const ChangingButton = observer(({ id, onClick, disabled, label }) => (
         {label}
     </button>
 ));
+
+export const InputField = observer(({id, inputVar, changeHandler, defaultValue, minLimit, maxLimit}) => {
+    const [inputValue, setInputValue] = useState(inputVar);
+
+    useEffect(() => {
+        setInputValue(inputVar);
+    }, [inputVar]);
+
+    const handleChange = (e) => {
+        let value = e.target.value;
+        // Remove leading zeros
+        if (/^0\d/.test(value)) {
+            value = value.replace(/^0+/, '');
+        }
+        // Limit the value to the allowable BPM range
+        const intValue = parseInt(value, 10);
+        if (intValue > maxLimit) {
+            value = maxLimit;
+        } else if (intValue < minLimit) {
+            value = minLimit;
+        }
+        setInputValue(value);
+    };
+
+    const handleBlur = () => {
+        const newValue = inputValue === '' ? defaultValue : Number(inputValue);
+        changeHandler(newValue);
+        setInputValue(newValue);
+    };
+
+    return (
+        <input
+            type="number"
+            id={id}
+            value={inputValue}
+            onChange={handleChange}
+            onBlur={handleBlur}
+            onKeyDown={(e) => preventNonDigitInput(e)}
+        />
+    );
+});
